\name{es_mef_sce}
\docType{data}
\alias{es_mef_sce}
\title{sample data for POWSC}
\description{
  This dataset is obtained from GEO under accession number GSE29087. It is generated to profile the transcriptomes for 92 single cells consisting of mouse embryonic fibroblast (MEF) and embryonic stem (ES) cells (Islam et al., 2011). The average sequencing depth for the dataset is around half a million.
}
\usage{es_mef_sce}
\format{A singlecellexperiment object contain the expressin data with two cell types}
\references{
  Islam, S., Kjällquist, U., Moliner, A., Zajac, P., Fan, J.-B., Lönnerberg, P., and
Linnarsson, S. (2011). Characterization of the single-cell transcriptional landscape
by highly multiplex RNA-seq. Genome research, 21(7), 1160–1167.
}
\keyword{datasets}
