% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDE.R
\name{runMAST}
\alias{runMAST}
\title{Run DE analysis by using MAST. Here we output two result tables corresponding to two forms of DE genes.
These parameters include four gene-wise parameters and two cell-wise parameters.}
\usage{
runMAST(sce)
}
\arguments{
\item{sce}{is a simulated scRNA-seq dataset with two-group conditions, e.g., treatment vs control.}
}
\value{
a list of three tables: the first table summaries the DE result for both forms of DE genes. cont table represents the result for continous case. disc table shows the result for discontinous case.
}
\description{
Run DE analysis by using MAST. Here we output two result tables corresponding to two forms of DE genes.
These parameters include four gene-wise parameters and two cell-wise parameters.
}
\examples{
de = runMAST(simData$sce)
}
