\name{sce}
\docType{data}
\alias{sce}
\title{sample data for GSE67835}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE67835}
}
\description{
  This dataset is obtained from GEO under accession number GSE67835. It includes 7 patients about 466 cells to capture the cellular complexity of the adult and fetal human brain at a whole transcriptome level. Healthy adult temporal lobe tissue was obtained from epileptic patients during temporal lobectomy for medically refractory seizures.
}
\usage{sce}
\format{A singlecellexperiment object contain the expressions data with three cell types about patient AS_7}
\references{
Darmanis, Spyros, et al. "A survey of human brain transcriptome diversity at the single cell level." Proceedings of the National Academy of Sciences 112.23 (2015): 7285-7290.
}
\keyword{datasets}
