% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPOWSC.R
\name{summary_POWSC}
\alias{summary_POWSC}
\title{summary of the result}
\usage{
summary_POWSC(POWSCobj, Form = c("I", "II"), Cell_Type = c("PW", "Multi"))
}
\arguments{
\item{POWSCobj}{a POWSC object from \code{runPOWSC}}

\item{Form}{choose from "I" or "II".}

\item{Cell_Type}{choose from "PW" or "Multi".}
}
\value{
return the summary of the power including stratified, marginal, and overall power.
}
\description{
summary of the result
}
\examples{
data("es_mef_sce")
sce = es_mef_sce[,colData(es_mef_sce)$cellTypes == "fibro"]
est_Paras = Est2Phase(sce)
sim_size = c(100, 400, 1000) # A numeric vector
pow_rslt = runPOWSC(sim_size = sim_size, est_Paras = est_Paras,per_DE=0.05, DE_Method = "MAST", Cell_Type = "PW") # Note, using our previous developed tool SC2P is faster.
summary_POWSC(pow_rslt, Form="II", Cell_Type = "PW")
}
