% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{SimulateMultiSCEs}
\alias{SimulateMultiSCEs}
\title{Simulate the data for multiple-group comparisons; e.g., different cell types in blood
It simulates the DE changes in two forms corresponding two types of DE genes}
\usage{
SimulateMultiSCEs(
  n = 1000,
  estParas_set,
  multiProb,
  delta1 = 0.1,
  delta2 = 0.5
)
}
\arguments{
\item{n}{the number of total cells for multiple groups; e.g., 1000, 2000, and etc.}

\item{estParas_set}{a set of parameters corresponding to different cell types.}

\item{multiProb}{a vector of probilities correponding to each cell type. It is not necessary to sum up to 1 because POWSC will normalize internally.}

\item{delta1}{the minimum of expression change used to determine the Form I DE.}

\item{delta2}{the minimum of log fold change used to determine the Form II DE.}
}
\value{
a list of simulated datasets. Each dataset corresponds to a pair-wise comparison including a series of metrics such as the DE gene indices for Form I and II DE genes, and simulated expression data in singlecellexperiment format.
}
\description{
Simulate the data for multiple-group comparisons; e.g., different cell types in blood
It simulates the DE changes in two forms corresponding two types of DE genes
}
\examples{
cell_per = c(0.4, 0.3, 0.2, 0.1)
sim = SimulateMultiSCEs(n = 6000, estParas_set = estParas_set, multiProb = cell_per)
}
